Expand `file` version constraints to include v6.
The Flutter repo and plugins are being migrated to NNBD.
The most recent changes (that I know of) is the file package
version `6.0.0-nullsafety.1`. If an app:

1. Depends on `flutter_cache_manager`, `flutter_driver` from sdk
2. Uses Flutter dev channel

Executing `pub get` will fail with something like this:

```
And because flutter_cache_manager >=1.2.0 depends on file ^5.1.0 and every version of flutter_driver from sdk depends on file 6.0.0-nullsafety.1, flutter_widget_from_html from path is incompatible with flutter_driver from sdk.
```

This commit expand the version constraints of the `file` dependency
to allow both v5 and v6. Testing seems to be unaffected but there are 
some [breaking changes][1] between file@5 and file@6 so it may be actually
unsafe to do this...

[1]: https://github.com/google/file.dart/blob/master/packages/file/CHANGELOG.md